requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
azure-identity==1.15.0
azure-keyvault-secrets==4.8.0
kubernetes==29.0.0
jinja2==3.1.3
python-dotenv==1.0.1
pydantic-settings==2.1.0
sqlalchemy==2.0.25
pyodbc==5.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6


